intel sycl introduces silent complex header loading chain. To adapt it, we need to do the same trick such that we can provide the implementation before loading complex header.
